Enable overriding and extending list value
With this commit, it's now possible to set a value to a specific item in
a list with the CLI: "kuku render ./templates --set a.0=foo"

Code is a bit tricky, prepare a pen if you want to proof-read it.
Otherwise, reading the unit tests should give a good confidence level.